new changes to assignment

+++ jsclasswork4.js
@@ -19,22 +19,24 @@
 //     }
 // ];
 users = JSON.parse(localStorage.getItem('users'));
-if(us)
+if(users == null) {
+    users = [];
+}
 
 let displayCard = () => {
     let showUser = '';
         for(let i = 0; i < users.length; i++) {
             showUser += `
-                <div>
-                    <img src="${users[i].profilePics}" alt="">
-                <div class="text">
-                    <p>Full Name : ${users[i].fullName}</p>
-                    <p>Email : ${users[i].email}</p>
-                    <p>Phone Number : ${users[i].phone}</p>
-                </div>
-                    <div class="card-buttons">
-                        <button id="deletebtn" class="deleteUser" onclick="deleteButton(${i})">Delete User</button>
-                        <button class="editUser">Edit User</button>
+                <div class="mx-auto">
+                    <img class="rounded img-fluid border border-light" style="width: 200px" src=" ${users[i].profilePics}" alt="">
+                    <div class="text-light mx-auto d-block">
+                        <p>Full Name : ${users[i].fullName}</p>
+                        <p>Email : ${users[i].email}</p>
+                        <p>Phone Number : ${users[i].phone}</p>
+                    </div>
+                    <div class="">
+                        <button id="deletebtn" class="btn btn-primary" onclick="deleteButton(${i})">Delete User</button>
+                        <button class="btn btn-primary" data-toggle="modal" data-target="#addButton" onclick=editButton(${i})>Edit User</button>
                     </div> 
                 </div> 
                 <hr>
@@ -45,10 +47,11 @@ let displayCard = () => {
 displayCard();
 
 let newUserInput = () => {
-    let newProfilePics = prompt('Enter Photo String');
-    let newFullName = prompt('Enter Full Name');
-    let newEmail = prompt('Enter Email');
-    let newPhoneNumber = prompt('Enter Phone Number');
+    addProfilePics = document.getElementById('profilePics').value;
+    let newProfilePics = addProfilePics.split('//')[2];
+    let newFullName = document.getElementById('fullName').value;
+    let newEmail = document.getElementById('email').value;
+    let newPhoneNumber = document.getElementById('phone').value;
 
     let newUser = {
         "profilePics" : newProfilePics,
@@ -58,20 +61,31 @@ let newUserInput = () => {
     }
 
     users.push(newUser);
+    alert('You have added a New User');
     localStorage.setItem('users', JSON.stringify(users));
-
+    location.reload();
 }
 
 
 
-function deleteButton(id) {
-    users.splice(id, 1);
-    localStorage.setItem('users', JSON.stringify(users));
-    displayCard();
+let deleteButton = (id) => {
+    let confirmation = confirm('Are you sure you want to delete');
+    if(confirmation = true) {
+        users.splice(id, 1);
+        localStorage.setItem('users', JSON.stringify(users));
+        displayCard();
+    }
+}
+
+let editButton = (id) => {
+    let userDetails = users[id];
+
+    document.getElementById('profilePics').value = "";
+    document.getElementById('fullName').value = userDetails.fullName;
+    document.getElementById('email').value = userDetails.email;
+    document.getElementById('phone').value = userDetails.phone;   
+}
+
+let update = (id) => {
     
 }
-
-function editButton(id) {
-    
-}
-
